requirements.txt
Flask
Flask-SQLAlchemy
mariadb
fabric